move index.html to task/docs
