+++ ARCHITECTURE.md
@@ -26,7 +26,7 @@ Legacy tests are found in the 'test' directory:
 
 The "tasks" map to top-level dbt commands. So `dbt run` => task.run.RunTask, etc. Some are more like abstract base classes (GraphRunnableTask, for example) but all the concrete types outside of task should map to tasks. Currently one executes at a time. The tasks kick off their “Runners” and those do execute in parallel. The parallelism is managed via a thread pool, in GraphRunnableTask.
 
-core/dbt/include/index.html
+core/dbt/task/docs/index.html
 This is the docs website code. It comes from the dbt-docs repository, and is generated when a release is packaged.
 
 ## Adapters
