+++ core/dbt/task/docs/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+DOCS_INDEX_FILE_PATH = os.path.normpath(os.path.join(os.path.dirname(__file__), "index.html"))
