+++ tests/functional/docs/test_static.py
@@ -1,7 +1,7 @@
 import pytest
 
 from dbt.clients.system import load_file_contents
-from dbt.task.generate import DOCS_INDEX_FILE_PATH
+from dbt.task.docs import DOCS_INDEX_FILE_PATH
 from dbt.tests.util import run_dbt
 import os
